Refactor code for better organization and separation
