+++ adaptivealerting/detectors/builders/base.py
@@ -1,0 +1,2 @@
+class DetectorBuilder:
+    """Base class for detectors builders."""
