+++ adaptivealerting/detectors/builders/exceptions.py
@@ -1,0 +1,4 @@
+from adaptivealerting import exceptions
+
+class DetectorBuilderError(exceptions.AdaptiveAlertingError):
+    """Base class for detector builder errors."""
