--- tests/test_detectors.py
@@ -1,8 +1,0 @@
-from adaptive_alerting_detector_build.detectors.base import BaseDetector
-
-
-def test_create_detector_():
-    def my_train_fn():
-        pass
-
-    MyDetector = BaseDetector(name="My First Detector", train=my_train_fn)
